Refactor code for base class
