--- client/app/cms/core/base.component.ts
@@ -1,6 +1,0 @@
-import { Input } from '@angular/core';
-import { PageBase } from './base.pagetype';
-
-export class ComponentBase<T extends PageBase> {
-    @Input() currentContent: T;
-}
