--- client/app/cms/core/base.pagetype.ts
@@ -1,5 +1,0 @@
-
-export class PageBase {
-    public pageName: string;
-    public linkUrl: string;
-}
