+++ client/app/cms/core/bases/base-component.ts
@@ -1,0 +1,6 @@
+import { Input } from '@angular/core';
+import { ContentData } from './content-data';
+
+export class BaseComponent<T extends ContentData> {
+    @Input() currentContent: T;
+}
