+++ client/app/cms/core/bases/content-data.ts
@@ -1,0 +1,4 @@
+
+export class ContentData {
+    public contentName: string;
+}
