+++ client/app/cms/core/bases/page-data.ts
@@ -1,0 +1,5 @@
+import { ContentData } from "./content-data";
+
+export class PageData extends ContentData {
+    public linkUrl: string;
+}
