+++ client/app/cms/core/decorators/content-type-metadata.ts
@@ -1,0 +1,5 @@
+export interface ContentTypeMetadata {
+    displayName?: string;
+    description?: string;
+    componentRef?: any;
+}
